# Conflicts:
#	app/src/test/java/io/gnosis/safe/ui/safe/SafeSelectionViewModelTest.kt
#	app/src/test/java/io/gnosis/safe/ui/safe/add/AddSafeNameViewModelTest.kt
#	app/src/test/java/io/gnosis/safe/ui/safe/add/AddSafeViewModelTest.kt
#	data/schemas/io.gnosis.data.db.HeimdallDatabase/5.json
#	data/src/main/java/io/gnosis/data/db/HeimdallDatabase.kt
#	data/src/main/java/io/gnosis/data/models/ChainInfo.kt
#	data/src/test/java/io/gnosis/data/repositories/ChainInfoRepositoryTest.kt
#	data/src/test/java/io/gnosis/data/repositories/SafeRepositoryTest.kt
#	data/src/test/java/io/gnosis/data/repositories/UnstoppableRepositoryTest.kt

+++ app/src/test/java/io/gnosis/safe/ui/assets/collectibles/CollectiblesViewModelTest.kt
@@ -59,7 +59,7 @@ class CollectiblesViewModelTest {
     fun `load - should emit collectibles view data list`() {
         val stateObserver = TestLiveDataObserver<BaseStateViewModel.State>()
         val chain = Chain(
-            BuildConfig.CHAIN_ID,
+            BuildConfig.CHAIN_ID.toBigInteger(),
             BuildConfig.BLOCKCHAIN_NAME,
             BuildConfig.CHAIN_TEXT_COLOR,
             BuildConfig.CHAIN_BACKGROUND_COLOR,
@@ -148,5 +148,4 @@ class CollectiblesViewModelTest {
             null
         )
     )
-
 }
