+++ data/src/main/java/io/gnosis/data/db/daos/ChainDao.kt
@@ -2,6 +2,7 @@ package io.gnosis.data.db.daos
 
 import androidx.room.*
 import io.gnosis.data.models.Chain
+import java.math.BigInteger
 
 @Dao
 interface ChainDao {
@@ -22,6 +23,6 @@ interface ChainDao {
     suspend fun loadByName(name: String): Chain?
 
     @Query("SELECT * FROM ${Chain.TABLE_NAME} WHERE ${Chain.COL_CHAIN_ID} = :id")
-    suspend fun loadByChainId(id: Int): Chain?
+    suspend fun loadByChainId(id: BigInteger): Chain?
 
 }
